requirements.txt
### Python 3.9.7 or higher
numpy
pykan
torch
scikit-learn
pandas 
tqdm